Fix edges for simulator screens.
On the simulator and DexLint the window for the pictures is bigger than
the actual pictures. Windows fills the void with the same color as the
upper-left pixel. In most cases, this pixel was white, so the pictures
look reasonable on the simulated displays. These pictures happened to
have a black edge which, in the simulator, effectively because a thick
black frame. To make the simulator look right I shrunk the images just
a hair. This should have a negligible effect on the GRIP hardware but a
visible effect on the simulator.