Changed name of picture file for GF tests.
There was a picture calle WaitSubjectReady but it showed a hand
pinching the manipulandum. It was only used in the scripts that run the
tests of the force sensors, so the picture was not incorrect. But the
filename could cause confusion, since it is not a general picture used
when waiting for the subject. So I just changed the name to be more
appropriate.

Note that this change was implemented after delivery of the scripts for
the FM2. The former filenames did not cause a functional problem with
the scripts, so there is no need to change on the FM2.